Wrong conditional statement
One of the status code check for response code was wrong. It was written as this:
```python
if policy_create_response.status_code == 200 or 210:
   # do stuff
```

This was causing an error. Even when the response code was a 500, the condition would match as "201" would always evaluate to true. I've corrected this statement with this code:

```python
if policy_create_response.status_code == 200 or policy_create_response==201:
   #do stuff
```